+++ src/pages/ranking/Ranking.tsx
@@ -6,16 +6,18 @@ import MenuBar from '@common/layout/MenuBar';
 import Header from '@common/layout/Header';
 import RankingContainer from '@features/ranking/ui/RankingContainer';
 import { RANKING_OPTIONS } from '@features/ranking/constants';
+import usePreloadImages from '@hooks/usePreloadImages';
+import { PRELOAD_IMAGES } from '@features/ranking/constants';
 import {
   useUserRankingQuery,
   useRankingPaginationQuery,
 } from '@features/ranking/queries';
 import useUserStore from '@/store/useUserStore';
-import { RankedUser } from '@/features/user/types';
 import { isLoggedIn } from '@features/user/service/authUtils';
 import Skeleton from '@common/layout/Skeleton';
-import usePreloadImages from '@/hooks/usePreloadImages';
-import { PRELOAD_IMAGES } from '@features/ranking/constants';
+
+import { generatePaginationPages } from '@utils/generatePaginationPages';
+import type { RankedUser } from '@/features/user/types';
 
 export default function Ranking() {
   const [currentPage, setCurrentPage] = useState<number>(1);
@@ -62,55 +64,11 @@ export default function Ranking() {
     setCurrentPage(page);
   };
 
-  const getPaginationPages = () => {
-    const MAX_VISIBLE = 7; // 최대 노출할 페이지 버튼 개수
-    const pages: (number | string)[] = [];
-
-    // 전체 페이지가 MAX_VISIBLE 이하라면 모두 표시
-    if (totalPage <= MAX_VISIBLE) {
-      for (let i = 1; i <= totalPage; i++) {
-        pages.push(i);
-      }
-      return pages;
-    }
-
-    // 최대 노출 버튼이 7개인 경우, 항상 첫 페이지와 마지막 페이지는 표시
-    // 가운데에 (MAX_VISIBLE - 2)개 버튼(여기선 5개)을 현재 페이지 중심으로 표시
-    const WINDOW_SIZE = MAX_VISIBLE - 2;
-
-    // 현재 페이지를 중심으로 시작 페이지 계산
-    let startPage = Math.max(2, currentPage - Math.floor(WINDOW_SIZE / 2));
-    let endPage = startPage + WINDOW_SIZE - 1;
-
-    // 만약 endPage가 마지막 전 페이지(totalPage - 1)보다 크다면 윈도우를 뒤로 당김
-    if (endPage > totalPage - 1) {
-      endPage = totalPage - 1;
-      startPage = endPage - WINDOW_SIZE + 1;
-    }
-
-    // 첫 페이지는 항상 추가
-    pages.push(1);
-
-    // 첫 페이지와 슬라이딩 윈도우 시작 사이에 공백이 있다면 ellipsis 추가
-    if (startPage > 2) {
-      pages.push('···');
-    }
-
-    // 슬라이딩 윈도우 페이지 추가
-    for (let i = startPage; i <= endPage; i++) {
-      pages.push(i);
-    }
-
-    // 슬라이딩 윈도우 끝과 마지막 페이지 사이에 공백이 있다면 ellipsis 추가
-    if (endPage < totalPage - 1) {
-      pages.push('···');
-    }
-
-    // 마지막 페이지 추가
-    pages.push(totalPage);
-
-    return pages;
-  };
+  const pages = generatePaginationPages({
+    currentPage,
+    totalPage,
+    maxVisible: 7,
+  });
 
   return (
     <>
@@ -159,7 +117,7 @@ export default function Ranking() {
           >
             ◀
           </S.RankingPaginationButton>
-          {getPaginationPages().map((page, index) =>
+          {pages.map((page, index) =>
             typeof page === 'number' ? (
               <S.RankingPaginationButton
                 key={index}
